% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_nhanes_study}
\alias{simulate_nhanes_study}
\title{Simulation study to evaluate imputer() vs. MICE using predictive mean matching}
\usage{
simulate_nhanes_study(
  numSims = 50,
  proportionNA = 0.2,
  nchains = 3,
  niter = 100
)
}
\arguments{
\item{numSims}{Number of simulations to run}

\item{proportionNA}{The proportion of all cells that should be NA,
across variables and observations, to be passed to \code{makeNA()}}

\item{nchains}{The number of chains (ie number of imputed datasets) to generate for imputer()}

\item{niter}{The number of iterations to complete in each chain (default is 100 for convergence) for imputer()}
}
\value{
A data.frame (\code{differences}) that includes the results of the simulation, with
a row for each simulation and a column for each variable
}
\description{
Completes a simulation study to compare the differences between \code{mice()} from the mice package
and \code{imputer()}, using an example complete case NHANES dataset from 2017-2018. Differences
are generated as the absolute difference between (1) \code{imputer} and the original NHANES data and (2)
\code{mice} and the original NHANES data, scaled by the difference \code{mice} and the original NHANES data.
Results are therefore interpreted as percent difference between \code{imputer} and \code{mice} as a function of
the difference between \code{mice} and the true cell values.
}
\examples{
# NOT RUN FOR RUNTIME
\dontrun{
library(imputevalR)
simulate_nhanes_study(numSims = 20, proportionNA = 0.2)
}

}
